Analysis.Context

Description

The context stores, among others, information obtained during verification, such as CFGs, invariants, etc. (see Context). Module VerificationReportInterface provides functions for obtaining and interfacing with a Context.

Documentation

data FContext

The context augmented with information on the current function

Constructors

FContext

Fields

f_ctxt :: Context

The context

f_entry :: Int

The entry address of the current function

f_init :: FInit

The initialization of the current function

mk_fcontext :: Context -> Int -> FContext

read_binary :: String -> String -> IO (Maybe Binary)

Reading a binary given a filename (ELF or MachO)

data CFG

A control flow graph with blocks and edges. A blockID (represented as an Int) is a unique identifier of a basic block. We store basic blocks twice: once as addresses, and once as instructions.

Constructors

CFG

Fields

cfg_blocks :: IntMap [Int]

A mapping of blockIDs to instruction addresses

cfg_edges :: IntMap IntSet

A mapping of blockIDs to sets of blocKIDs

cfg_addr_to_blockID :: IntMap Int

A mapping of instruction addresses to blockIDs

cfg_fresh :: Int

A fresh blockID

cfg_instrs :: IntMap [Instruction]

A mapping of blockIDs to lists of disassembled instructions.

data JumpTable

A jump table with : index: an operand containing a bounded index at the beginning of execution of the block bound: the bound on idx trgt: the operand containg the jump target at the end of execution of the block tbl: a table from values of idx to resulting jump targets

Constructors

JumpTable

Fields

jtbl_index :: Operand

jtbl_bound :: Int

jtbl_target :: Operand

jtbl_table :: IntMap Word64

data Indirection

An indirection is either a jump table or a set of resolved jump targets

Constructors

Indirection_JumpTable JumpTable

Indirection_Resolved (Set ResolvedJumpTarget)

Indirection_Unresolved

type Indirections = IntMap Indirection

Per instruction address, a set of possibly resolved jump targets

data VerificationResult

An enumeration indicating the result of verification over a function

Constructors

VerificationSuccess
Function was succesfully verified

VerificationSuccesWithAssumptions
Function was succesfully verified, but required assertions

VerificationUnresolvedIndirection
Function contains an unresolved indirection

VerificationError String
There was some verification error, e.g., return adresss overwrite

Unverified
The function has not been verified.

type Predicate = Sstate SPointer

Predicates: symbolic states

type Invariants = IntMap Predicate

Invariants: a mapping of blockIDs to predicates

data NodeInfo

For each leaf-node in a CFG we store the following info. (TODO MOVE?)

Constructors

Normal
The basic block behaves normally, e.g., a ret

UnresolvedIndirection
The basic block ends in an unresolved indirection

Terminal
The basic blocks ends with, e.g., a call to exit()

type Postconditions = Set (NodeInfo, Predicate)

Postconditions: for each final block the NodeInfo and the final predicate after execution of the block

data PointerDomain

An abstract domain for pointers

Constructors

Domain_Bases (NESet PointerBase)

Domain_Sources (NESet BotSrc)

type FInit = Map Register SPointer

A function initialisation consists of a mapping of registers to symbolic pointers.

data FReturnBehavior

A function call

Constructors

Terminating
The function does never return

ReturningWith Predicate
The function returns withg the symbolic changes stored in the predicate

UnknownRetBehavior
It is unknown whether the function returns or not

data SStatePart

A symbolic state part

Constructors

SSP_Reg Register
A register

SSP_Mem SPointer Int

data Context_

The context datastructure.

S: Information Statically obtained by reading from the binary

D: Information Dynamically updated during verification

Constructors

Context_

Fields

ctxt__config :: !Config

S: the configuration file in ./config

ctxt__verbose :: !Bool

S: is verbose output requested?

ctxt__syms :: !SymbolTable

S: the symbol table: a mapping of addresses to function names for external functions

ctxt__sections :: !SectionsInfo

S: information on segments/section

ctxt__relocs :: !(Set Relocation)

S: relocation entries

ctxt__dirname :: !String

S: the name of the directory where the .dump, .entry, .sections and .symbols files reside

ctxt__name :: !String

S: the name of the binary

ctxt__start :: !Int

S: the address of the _start symbol

ctxt__entries :: !Graph

D: a graph with an edge (e0,e1) if entry address e0 calls entry address e1, and e0 and e1 have not been verified yet

ctxt__cfgs :: !(IntMap CFG)

D: the currently known control flow graphs per function entry

ctxt__calls :: !(IntMap FReturnBehavior)

D: the currently known and verified entry addresses of functions mapped to return-information

ctxt__invs :: !(IntMap Invariants)

D: the currently known invariants

ctxt__posts :: !(IntMap Postconditions)

D: the currently known postconditions

ctxt__stateparts :: !(IntMap (Set SStatePart))

D: the state parts read from/written to be the function

ctxt__inds :: !Indirections

D: the currently known indirections

ctxt__finits :: !(IntMap FInit)

D: the currently known function initialisations

ctxt__vcs :: !(IntMap VCS)

D: the verification conditions

ctxt__results :: !(IntMap VerificationResult)

D: the verification result

ctxt__recursions :: !(IntMap IntSet)

D: a mapping from function entries to the set of mutually recursive functions entries they occur in

ctxt__runningtime :: !Int64

D: running time of lifting effort

data Context

Constructors

Context

Fields

ctxt_binary :: !Binary

ctxt_ioref :: IORef (IntMap Instruction)

ctxt_ctxt_ :: !Context_

ctxt_config :: Context -> Config

ctxt_verbose :: Context -> Bool

ctxt_syms :: Context -> SymbolTable

ctxt_sections :: Context -> SectionsInfo

ctxt_relocs :: Context -> Set Relocation

ctxt_dirname :: Context -> String

ctxt_name :: Context -> String

ctxt_start :: Context -> Int

ctxt_entries :: Context -> Graph

ctxt_cfgs :: Context -> IntMap CFG

ctxt_calls :: Context -> IntMap FReturnBehavior

ctxt_invs :: Context -> IntMap Invariants

ctxt_posts :: Context -> IntMap Postconditions

ctxt_stateparts :: Context -> IntMap (Set SStatePart)

ctxt_inds :: Context -> Indirections

ctxt_finits :: Context -> IntMap FInit

ctxt_vcs :: Context -> IntMap VCS

ctxt_results :: Context -> IntMap VerificationResult

ctxt_recursions :: Context -> IntMap IntSet

ctxt_runningtime :: Context -> Int64

set_ctxt_syms :: SymbolTable -> Context -> Context

set_ctxt_sections :: SectionsInfo -> Context -> Context

set_ctxt_relocs :: Set Relocation -> Context -> Context

set_ctxt_start :: Int -> Context -> Context

set_ctxt_entries :: Graph -> Context -> Context

set_ctxt_calls :: IntMap FReturnBehavior -> Context -> Context

set_ctxt_inds :: Indirections -> Context -> Context

set_ctxt_posts :: IntMap Postconditions -> Context -> Context

set_ctxt_stateparts :: IntMap (Set SStatePart) -> Context -> Context

set_ctxt_vcs :: IntMap VCS -> Context -> Context

set_ctxt_finits :: IntMap FInit -> Context -> Context

set_ctxt_invs :: IntMap Invariants -> Context -> Context

set_ctxt_cfgs :: IntMap CFG -> Context -> Context

set_ctxt_results :: IntMap VerificationResult -> Context -> Context

set_ctxt_recursions :: IntMap IntSet -> Context -> Context

set_ctxt_runningtime :: Int64 -> Context -> Context

init_context :: Binary -> IORef (IntMap Instruction) -> Config -> Bool -> String -> String -> Context

intialize an empty context based on the command-line parameters

purge_context :: Context -> Context_

purge the context before exporting it (may save a lot of disk space)

ctxt_symbol_table :: Context -> IntMap Symbol

Getting the symbol table

read_from_ro_datasection

Arguments

:: Context
The context

-> Word64
An address

-> Int
Size, i.e., the number of bytes to read

-> Maybe Word64

Reading from a read-only data section.

Reads maximally up to 8 bytes. Returns Nothing if the given address is out-of-range.

read_from_datasection

Arguments

:: Context
The context

-> Word64
An address

-> Int
Size, i.e., the number of bytes to read

-> Maybe Word64

Reading from a writable data section.

Reads maximally up to 8 bytes. Returns Nothing if the given address is out-of-range.

is_roughly_an_address

Arguments

:: Context
The context

-> Word64
An address

-> Bool

Is the immediate roughly in range to be an address?

find_section_for_address

Arguments

:: Context
The context

-> Word64
An address

-> Maybe (String, String, Word64, Word64)

Find a section for an address (see SectionsInfo)

fetch_instruction

Arguments

:: Context
The context

-> Word64
An address

-> IO (Maybe Instruction)

Fetching an instruction

Returns Nothing if the given address is out-of-range.

pp_instruction

Arguments

:: Context
The context

-> Instruction
An instruction

-> String

Pretty printing an instruction

show_finit :: FInit -> String

Show function initialisaExpr

is_assertion :: VerificationCondition -> Bool

Is the given verification condition an assertion?

is_precondition :: VerificationCondition -> Bool

Is the given verification condition a precondition?

is_func_constraint :: VerificationCondition -> Bool

Is the given verification condition a function constraint?

is_functionpointers :: VerificationCondition -> Bool

Is the given verification condition a function pointer introduction?

count_instructions_with_assertions :: Set VerificationCondition -> Int

Count the number of assertions in the set of verification conditions.

ctxt_continue_on_unknown_instruction :: Context -> Bool

ctxt_generate_pdfs :: Context -> Bool

ctxt_verbose_logs :: Context -> Bool

ctxt_store_preconditions_in_L0 :: Context -> Bool

ctxt_store_assertions_in_L0 :: Context -> Bool

ctxt_max_time :: Context -> Int

ctxt_max_num_of_cases :: Context -> Int

ctxt_max_num_of_bases :: Context -> Int

ctxt_max_num_of_sources :: Context -> Int

ctxt_max_jump_table_size :: Context -> Int

ctxt_max_expr_size :: Context -> Int

ctxt_read_L0

Arguments

:: String
The directory name

-> String
The file name of the binary

-> IO Context

Read in the .L0 file from a file with the given file name. May produce an error if no L0 can be read from the file. Returns the L0 stored in the .L0 file.

address_has_instruction :: Integral a => Context -> a -> Bool

Returns true iff an instruction can be fetched from the address.

Orphan instances

Show VerificationCondition

Instance details

Methods

showsPrec :: Int -> VerificationCondition -> ShowS

show :: VerificationCondition -> String

showList :: [VerificationCondition] -> ShowS

Serialize ResolvedJumpTarget

Instance details

Methods

put :: Putter ResolvedJumpTarget

get :: Get ResolvedJumpTarget

Serialize Relocation

Instance details

Methods

put :: Putter Relocation

get :: Get Relocation

Serialize SymbolTable

Instance details

Methods

put :: Putter SymbolTable

get :: Get SymbolTable

Serialize Symbol

Instance details

Methods

put :: Putter Symbol

get :: Get Symbol

Serialize SectionsInfo

Instance details

Methods

put :: Putter SectionsInfo

get :: Get SectionsInfo

Serialize VerificationCondition

Instance details

Methods

put :: Putter VerificationCondition

get :: Get VerificationCondition

NFData ResolvedJumpTarget

Instance details

Methods

rnf :: ResolvedJumpTarget -> ()

NFData Relocation

Instance details

Methods

rnf :: Relocation -> ()

NFData SymbolTable

Instance details

Methods

rnf :: SymbolTable -> ()

NFData Symbol

Instance details

Methods

rnf :: Symbol -> ()

NFData SectionsInfo

Instance details

Methods

rnf :: SectionsInfo -> ()

NFData VerificationCondition

Instance details

Methods

rnf :: VerificationCondition -> ()