Base

Description

Documentation

showHex :: (Integral a, Show a) => a -> String

Show the integer in hex.

showHex_list :: (Integral a, Show a) => [a] -> [Char]

Show an integer list as hex-list.

showHex_set :: IntSet -> [Char]

Show an integer set as hex-list.

showHex_option :: (Integral a, Show a) => Maybe a -> [Char]

Show an optional integer as an optional hex.

readHex' :: (Eq a, Num a) => String -> a

Read an int from a string storing a hex.

im_lookup :: [Char] -> IntMap p -> Key -> p

Lookup and produce error message if key does not exists in map.

orElse :: Maybe p -> p -> p

use a default value in case of Nothing

existsAndSatisfies :: Maybe t -> (t -> Bool) -> Bool

A value exists (is not Nothing) and satisfies the predicate

pair :: a -> b -> (a, b)

create a tuple

findString :: Eq a => [a] -> [a] -> Maybe Int

Find the index of one string in another.

takeUntilString :: String -> String -> String

Take until the occurrence of the string

strip_parentheses :: [Char] -> [Char]

Strip outer parentheses from a string, if it has them.

partitionWith :: (a -> Either b c) -> [a] -> ([b], [c])

Uses a function to determine which of two output lists an input element should join

bytes_to_word :: [Word8] -> Word64

In little endian, convert a byte-list to a 64 bit word. Assume the list is at most length 8.

word_to_sint :: Int -> Word64 -> Int

Convert first n bytes of a word to an integer. Assume n<8.

average :: (Real a, Fractional b) => [a] -> b

average of list of numbers

crossProduct :: [[a]] -> [[a]]

crossProduct_size :: Foldable t => [t a] -> Int

neFromList :: Ord a => [a] -> NESet a

neSetToList :: NESet a -> [a]

data Graph

Generic graph with ints as vertices.

Constructors

Edges (IntMap IntSet)

graph_add_edges :: Graph -> Key -> IntSet -> Graph

add edges from v to all vertices vs

graph_delete :: Graph -> Key -> Graph

delete all edges with v as parent or child

graph_is_parent :: Graph -> Key -> Bool

is v parent of an edge?

graph_is_vertex :: Graph -> Key -> Bool

is v a vertex in the graph?

graph_is_edge :: Graph -> Key -> Key -> Bool

is (v0,v1) an edge?

graph_nontrivial_scc :: Graph -> IntSet

retrieve a non-trivial SCC, if any exists

graph_find_next :: Graph -> Maybe Int

find next vertex to consider: either a terminal vertex (if any) or the head of an SCC

putNESetOf :: Putter a -> Putter (NESet a)

getNESetOf :: Ord a => Get a -> Get (NESet a)

Read as a list of elements.

hex_color_of_text :: String -> String

decide whether text should be white or black based on background color

hex_colors :: [[Char]]

A list of RGB colors

allp :: [a -> Bool] -> a -> Bool

Orphan instances

(Ord a, Serialize a) => Serialize (NESet a)

Instance details

Methods

put :: Putter (NESet a)

get :: Get (NESet a)