Binary.Generic

Documentation

data SectionsInfo

Information on the sections in the binary

Constructors

SectionsInfo

Fields

si_sections :: ![(String, String, Word64, Word64, Word64)]

Sections: segment names, section names, addresses, sizes, and alignment.

si_min_address :: !Word64

si_max_address :: !Word64

data SymbolTable

Constructors

SymbolTable

Fields

symboltable_symbols :: !(IntMap Symbol)

symboltable_exterbals :: !(Set String)

data Relocation

Constructors

Relocation Word64 Word64
8: At address a0, address a1 has been written, i.e., qword ptr[a0] == a1

class BinaryClass a where

Methods

binary_read_bytestring :: a -> Word64 -> Int -> Maybe ByteString

binary_read_ro_data :: a -> Word64 -> Int -> Maybe [Word8]

binary_read_data :: a -> Word64 -> Int -> Maybe [Word8]

binary_get_sections_info :: a -> SectionsInfo

binary_get_symbols :: a -> SymbolTable

binary_get_relocations :: a -> Set Relocation

binary_pp :: a -> String

binary_entry :: a -> Word64

binary_text_section_size :: a -> Int

binary_dir_name :: a -> String

binary_file_name :: a -> String

data Binary

Constructors

forall b.BinaryClass b => Binary b

binary_get_symbol_table :: BinaryClass a => a -> IntMap Symbol

binary_get_global_symbols :: BinaryClass a => a -> Set String

symbol_to_name :: Symbol -> String

is_roughly_an_address

Arguments

:: BinaryClass bin

=> bin
The binary

-> Word64
An address

-> Bool

Is the immediate roughly in range to be an address?

find_section_for_address

Arguments

:: BinaryClass bin

=> bin
The binary

-> Word64
An address

-> Maybe (String, String, Word64, Word64, Word64)

Find a section for an address (see SectionsInfo)

read_from_ro_datasection

Arguments

:: BinaryClass bin

=> bin
The binary

-> Word64
An address

-> Int
Size, i.e., the number of bytes to read

-> Maybe Word64

Reading from a read-only data section.

Reads maximally up to 8 bytes. Returns Nothing if the given address is out-of-range.

find_section_ending_at

Arguments

:: BinaryClass bin

=> bin
The binary

-> Word64
An address

-> Maybe (String, String, Word64, Word64, Word64)

Find a section ending at address (see SectionsInfo)

fetch_instruction

Arguments

:: BinaryClass bin

=> bin

-> Word64
An address

-> Maybe Instruction

Fetching an instruction Returns Nothing if the given address is out-of-range.

address_has_instruction

Arguments

:: BinaryClass bin

=> bin

-> Word64
An address

-> Bool

Returns true iff an instruction can be fetched from the address.