CFG_Gen

Description

Documentation

cfg_gen

Arguments

:: Context
The context

-> Int
The entry point of the function

-> IO (Set (Instr, Int), CFG)

Produce a CFG

Given the entry point of the function, generate either a CFG, or a set of new entry points to be analyzed first. The set of new entry points are function entries called by the current function, but for which we do not know yet whether they terminate or not. If a CFG is returned, then all function calls in that CFG have already been analyzed.

cfg_to_dot

Arguments

:: Context
The context

-> CFG
The CFG

-> String

Export a CFG to .dot file

Strongly connected components get the same color.

is_end_node

Arguments

:: CFG
The CFG

-> Int
The blockID

-> Bool

Returns true if the given blockID is a leaf-node in the given CFG.

node_info_of :: Context -> CFG -> Int -> NodeInfo

Returns the NodeInfo of a given blockID.

Assumes the given blockID corresponds to a leaf-node.

stepA

Arguments

:: Context
The context

-> Int
The entry address

-> Int
The instruction address

-> IO (Either (Set (Instr, Int)) [(Int, Bool)])

An abstract step function

Given the entry address of the function currently under investigation, and the instruction address of the current instruction, try to get the set of next instruction addresses.

This returns either: * a set of tuples (i,a) where i is an instruction and a its address. All these instructions are function calls that need to be analyzed before this current function entry can continue. * a list of tuples (a,b) where a is an instruction address that may follow the current instruction, and b is a Bool indicating whether that address belongs to a call