Context

Description

A Context stores all the information retrieved from the binary, as well as the command-line parameters passed to FoxDec. Moreover, it stores information obtained during verification, such as CFGs, invariants, etc. (see Context).

Module VerificationReportInterface provides functions for obtaining and interfacing with a Context.

Documentation

data CFG

A control flow graph with blocks and edges. A blockID (represented as an Int) is a unique identifier of a basic block. We store basic blocks twice: once as addresses, and once as instructions.

Constructors

CFG

Fields

cfg_blocks :: IntMap [Int]

A mapping of blockIDs to instruction addresses

cfg_edges :: IntMap IntSet

A mapping of blockIDs to sets of blocKIDs

cfg_addr_to_blockID :: IntMap Int

A mapping of instruction addresses to blockIDs

cfg_fresh :: Int

A fresh blockID

cfg_instrs :: IntMap [Instr]

A mapping of blockIDs to lists of disassembled instructions.

type Indirections = IntMap IntSet

Per instruction address, a set of jump targets.

type SectionsInfo = [(String, String, Int, Int)]

Sections: segment names, section names, addresses and sizes.

data VerificationResult

An enumeration indicating the result of verification over a function

Constructors

VerificationSuccess
Function was succesfully verified

VerificationSuccesWithAssertions
Function was succesfully verified, but required assertions

VerificationUnresolvedIndirection
Function contains an unresolved indirection

VerificationError
There was some verification error, e.g., return adresss overwrite

Unverified
The function has not been verified.

type Invariants = IntMap Pred

Invariants: a mapping of blockIDs to predicates

data NodeInfo

For each leaf-node in a CFG we store the following info.

Constructors

Normal
The basic block behaves normally, e.g., a ret

UnresolvedIndirection
The basic block ends in an unresolved indirection

Terminal
The basic blocks ends with, e.g., a call to exit()

data Report

Per function, we report on:

Constructors

Report

Fields

report_cfg :: CFG

The control flow graph

report_invs :: Invariants

The invariants

report_posts :: Set (NodeInfo, Pred)

The postcondition(s)

report_result :: VerificationResult

The verification result

report_vcs :: Set VerificationCondition

The verification conditions

data Context

The context datastructure.

S: Information Statically obtained by reading from the binary

D: Information Dynamically updated during verification

Constructors

Context

Fields

ctxt_dump :: IntMap Word8

S: mapping from addresses to bytes (data and instructions from the binary/executable)

ctxt_syms :: IntMap String

S: the symbol table: a mapping of addresses to function names for external functions

ctxt_calls :: IntMap Bool

D: the currently known and verified entry addresses of functions (true iff always terminating, i.e., non-returning)

ctxt_entries :: Graph

D: a graph with an edge (e0,e1) if entry address e0 calls entry address e1, and e0 and e1 have not been verified yet

ctxt_sections :: SectionsInfo

S: information on segments/section

ctxt_dirname :: String

S: the name of the directory where the .dump, .entry, .sections and .symbols files reside

ctxt_name :: String

S: the name of the binary

ctxt_generate_pdfs :: Bool

S: do we call graphviz to generate PDFs from .dot files?

ctxt_inds :: Indirections

D: the currently known indirections

ctxt_report :: IntMap Report

D: a mapping from function entries to reports storing verification results

read_from_datasection

Arguments

:: Context
The context

-> Word64
An address

-> Int
Size, i.e., the number of bytes to read

-> Maybe Word64

Reading from a data section.

Reads maximally up to 8 bytes. Returns Nothing if the given address is out-of-range.

find_section_for_address

Arguments

:: Context
The context

-> Int
An address

-> Maybe (String, String, Int, Int)

Find a section for an address (see SectionsInfo)

fetch_instruction

Arguments

:: Context
The context

-> Int
An address

-> IO (Maybe Instr)

Fetching an instruction

Returns Nothing if the given address is out-of-range.

pp_instruction

Arguments

:: Context
The context

-> Instr
An instruction

-> String

Pretty printing an instruction