ControlFlow

Description

Contains function relating to control flow, including functions for resolving the targets of jumps and calls.

Documentation

data ResolvedJumpTarget

Resolving the operand of a jump/call can produce one of the following.

Constructors

Unresolved
An indirect branch that has not been resolved yet

External String
A call to external function f

ImmediateAddress Word64
An internal call to the given address

post :: CFG -> Key -> IntSet

The set of next blocks from the given block

fetch_block

Arguments

:: CFG
The CFG

-> Int
The blockID

-> [Instr]

Fetching an instruction list given a block ID

address_has_instruction :: Integral a => Context -> a -> Bool

Returns true iff an instruction can be fetched from the address.

address_has_symbol :: Integral a => Context -> a -> Bool

Returns true iff a symbol is associated with the address.

address_is_external :: Integral a => Context -> a -> Bool

Returns truee if the adress is external, i.e., has no instruction or has a symbol.

operand_static_resolve

Arguments

:: Context
The context

-> Instr
The instruction

-> Maybe Operand
The operand of the instruction to be resolved

-> ResolvedJumpTarget

many operands can statically be resolved, even though technically they are indirect (relative to RIP). Examples:

10005464e: call RIP + 1751660 resolves to an immediate jump target by resolving the RIP-relative addressing.

10005464e: call qword ptr [RIP + 1751660] read from address 1002000c0, but address has a symbol associated to it. This function call will resolve to an external function.

resolve_jump_target

Arguments

:: Context
The context

-> Instr
The instruction

-> [ResolvedJumpTarget]

Resolves the first operand of a call or jump instruction. First tries to see if the instruction is an indirection, that has already been resolved. If not, try to statically resolve the first operand using operand_static_resolve. If that resolves to an external symbol libc_start_main, then this call is actually an indirection (and thus currently unresolved). If that resolves to an immediate value, see if that immediate value corresponds to an external function or an internal function.

Returns a list of ResolvedJumpTarget, since an indirection may be resolved to multiple targets.

get_internal_addresses

Arguments

:: ResolvedJumpTarget
A resolved jump target

-> [Int]

Given a resolved jump target, get a possibly empty list of internal addresses to which the jump target can jump.

instruction_jumps_to_external

Arguments

:: Context
The context

-> Instr
The instruction

-> Bool

Returns true iff the instruction resolves to external targets only.

show_block

Arguments

:: CFG
The CFG

-> Int
The blockID

-> String

Shows the block associated to the givern blockID.

show_invariants

Arguments

:: CFG
The CFG

-> Invariants
The invariants

-> String

Shows invariants.

function_name_of_entry

Arguments

:: Context
The context

-> Int
The entry address

-> String

Tries to retrieve a function name with an entry address. If the entry matches a known symbol, return that. Otherwise, simply return the entry address itself in hexadecimal notation. However, there is one exception: if the first instruction at the entry address immediately jumps to an external function, return the name of that external function instead. This happens in a .got section.

function_name_of_instruction

Arguments

:: Context
The context

-> Instr
The instruction

-> String

Tries to retrieve a function name for a call-instruction (see function_name_of_entry).

Returns the empty string if the given instruction is not a call or a jump.