Conventions

Description

In this file, we eumerate assumptions made on calling conventions and over external functions.

Documentation

exiting_functon_calls :: [[Char]]

A list of function names of functions that never return.

sections_modifiable_by_external_functions :: [a]

An overview of which sections can be modified by external functions. Other sections are assumed NOT to be modified by external functions.

sections_with_instructions :: [([Char], [Char])]

Overview of sections with instructions.

section_is_unwritable :: ([Char], [Char]) -> Bool

Sections in the following list are assumed not to be modifiable during execution, i.e., constant.

callee_saved_registers :: [Register]

A list if registers that are non-volatile, i.e., that must be preserved by a function (callee-saved)

return_registers :: [Register]

A list of registers that may be used for return values

parameter_registers :: [Register]

A list of registers used as parameters

max_jump_table_size :: Integer

An overapproximation of the maximum number of entries in a jump table. Does not affect soundness, but if the value is set too low, then more indirections may be left unresolved.