Data.Indirection

Documentation

data JumpTable

A jump table with : index: an operand containing a bounded index at the beginning of execution of the block bound: the bound on idx trgt: the operand containg the jump target at the end of execution of the block tbl: a table from values of idx to resulting jump targets

Constructors

JumpTable

Fields

jtbl_index :: Operand

jtbl_bound :: Int

jtbl_target :: Operand

jtbl_table :: IntMap Word64

data Indirection

An indirection is either a jump table or a set of resolved jump targets

Constructors

Indirection_JumpTable JumpTable

Indirection_Resolved ResolvedJumpTarget

Indirection_Unresolved

type Indirections = Set Indirection