Data.JSON

Description

Taxonomy

data Address

Address

Constructors

AddressRegister Register
Reading a pointer from a storage

AddressImm Word64
Immediate value

AddressPlus Address Address
Plus

AddressMinus Address Address
Minus

AddressTimes Address Address
Times

data Operand

Operand

Constructors

Memory Address Int
A region in memory, whose address is stored in the given state part and whose size in bytes is given in the Int

EffectiveAddress Address
An address itself, but not the value stored at the address.

Reg Register
A storage location such as a register or a variable

Immediate Word64
An immediate value

data Instruction

Instruction

Constructors

Instruction

Fields

addr :: Word64

address

prefix :: Maybe Prefix

prefix, e.g., lock or repz

opcode :: Opcode

opcode/mnemonic

dest :: Maybe Operand

- ^ optional: destination operand

srcs :: [Operand]

optional: operand

size :: Int

size of instruction