Data.SValue

Description

Documentation

data PtrOffset

A `pointer value` may have a `pointer offset` that is either a known fixed immediate or unknown

Constructors

PtrOffset Word64

UnknownOffset

data SPointer

A `pointer value` consists of a `pointer base` and a `pointer offset`

Constructors

Base_StackPointer String PtrOffset
The stackpointer of the given function

Base_Immediate Word64 PtrOffset
An immediate pointer

Base_Malloc (Maybe Word64) (Maybe String) PtrOffset
A malloc return value

Base_FunctionPtr Word64 String PtrOffset
A function pointer (external)

Base_ReturnAddr String
The return address of the given function

Base_TLS PtrOffset
The Thread Local Storage

Base_StatePart StatePart PtrOffset
The value initially stored in some statepart

Base_FunctionReturn String PtrOffset
The return value of a function

data SAddend

A `pointer addend` consists of a `pointer base`, some constant relativ to the initial state, or the return value of some function.

Constructors

SAddend_StackPointer String

SAddend_Immediate Word64

SAddend_Malloc (Maybe Word64) (Maybe String)

SAddend_FunctionPtr Word64 String

SAddend_ReturnAddr String

SAddend_TLS

SAddend_StatePart StatePart

SAddend_FunctionReturn String

data SValue

Constructors

SConcrete (NESet SimpleExpr)

SAddends (NESet (NESet SAddend))

Top

show_set :: [[Char]] -> [Char]

isImmediateBase :: SAddend -> Bool

isConcrete :: SValue -> Bool

has_unknown_offset :: SPointer -> Bool

liftOffsetMod :: (Word64 -> Word64) -> PtrOffset -> PtrOffset

mod_offset :: (Word64 -> Word64) -> SPointer -> SPointer