Data.VerificationCondition

Documentation

data PointerAnalysisResult v

Constructors

PointerAnalysisResult

Fields

pa_mem_write :: Maybe v

pa_mem_reads :: [Maybe v]

show_pars :: Show a => [Maybe a] -> [Char]

data VerificationCondition v

Constructors

FunctionPointers Word64 IntSet
A set of function pointers passed to a function

PointerAnalysis Word64 (PointerAnalysisResult v)
A pointer analysis result for the instruction at the address

type VCS v = Set (VerificationCondition v)

gather_pa_results :: VCS v -> IntMap (PointerAnalysisResult v)