Data.X86.Instruction

Documentation

data Instruction

Constructors

Instruction

Fields

inAddress :: Word64

inPrefix :: [Prefix]

inOperation :: Opcode

inDest :: Maybe Operand

inOperands :: [Operand]

inSize :: Int

data Prefix

Constructors

PrefixO16

PrefixA32

PrefixRepNE

PrefixRep

PrefixLock

PrefixSeg SReg

PrefixRex Word8

data Operand

Constructors

Op_Mem

Fields

mSize :: BitSize

mASize :: BitSize

mReg :: Register

mIdx :: Register

mScale :: Word8

mDisp :: Int

mSeg :: Maybe SReg

Op_Reg Register

Op_Imm Immediate

Op_Jmp Immediate

Op_Const Int

Op_Near Operand

Op_Far Operand

data Immediate

Constructors

Immediate

Fields

iSize :: BitSize

iValue :: Word64

operand_size :: Operand -> ByteSize

mk_RSP_mem_operand :: ByteSize -> Operand

srcs :: Instruction -> [Operand]

canonicalize :: Instruction -> [Instruction]

Canonicalizes an instruction by making sure there is at most one output and inputs/outputs are syntactically clear. | For example, the IMUL instruction comes in different flavors that hide the actual inputs and outputs to multiple registers

canonicalize_sextend1 :: Instruction -> [Instruction]

canonicalize_sextend2 :: Instruction -> [Instruction]

canonicalize_mul :: Instruction -> [Instruction]

canonicalize_div :: Instruction -> [Instruction]

mnemonic_reads_from_all_operands :: Opcode -> Bool

mnemonic_reads_from_all_but_first_operands :: Opcode -> Bool

remove_destination :: Opcode -> Bool

do_not_modify :: Opcode -> Bool

lowpart :: Opcode -> Opcode

hipart :: Opcode -> Opcode