Generic.Operand

Documentation

data GenericOperand storage

A generic statepart, based on polymorphic type storage.

Constructors

Memory (GenericAddress storage) Int
A region in memory, whose address is stored in the given state part and whose size in bytes is given in the Int

EffectiveAddress (GenericAddress storage)
An address itself, but not the value stored at the address.

Storage storage
A storage location such as a register or a variable

Immediate Word64
An immediate value