Generic.SymbolicConstituents

Documentation

data VerificationCondition

A verification condition is either: * A precondition of the form: > Precondition (a0,si0) (a1,si1) This formulates that at the initial state the two regions must be separate. * An assertion of the form: > Assertion a (a0,si0) (a1,si1) This formulates that dynamically, whenever address a is executed, the two regions are asserted to be separate. * A function constraint of the form: > FunctionConstraint foo [(RDI, v0), (RSI, v1), ...] { sp0,sp1,... } This formulates that a function call to function foo with values v0, v1, ... stored in the registers should not overwrite certain state parts.

Constructors

Precondition SimpleExpr Int SimpleExpr Int
Precondition: lhs SEP rhs

Assertion SimpleExpr SimpleExpr Int SimpleExpr Int
Assertion: @address, lhs SEP rhs

FunctionConstraint String Word64 [(Register, SimpleExpr)] (Set StatePart)
Function name, address, of call, with param registers

FunctionPointers Word64 IntSet
A set of function pointers passed to a function

type VCS = Set VerificationCondition

An acornym for a set of verification conditions

add_function_pointer :: Word64 -> Key -> (a, Set VerificationCondition) -> (a, Set VerificationCondition)

Add a function_pointer_intro to the given symbolic predicate

data SymbolicOperation v

Constructors

SO_Op Opcode Int (Maybe Int) [v]

SO_Bit Int v

SO_SExtend Int Int v

SO_Overwrite Int v v

SO_Plus v v

SO_Minus v v

SO_Times v v

data RegionSize

Constructors

Nat Word64

UnknownSize

class (Ord v, Eq v, Show v, Eq p, Ord p, Show p) => SymbolicExecutable ctxt v p | ctxt -> v p where

Methods

sseparate :: ctxt -> String -> p -> RegionSize -> p -> RegionSize -> Bool

senclosed :: ctxt -> p -> RegionSize -> p -> RegionSize -> Bool

salias :: ctxt -> p -> RegionSize -> p -> RegionSize -> Bool

ssensitive :: ctxt -> p -> RegionSize -> v -> Bool

sread_from_ro_data :: ctxt -> p -> RegionSize -> Maybe v

smk_mem_addresses :: ctxt -> String -> v -> Set p

sjoin_values :: Foldable t => ctxt -> String -> t v -> v

swiden_values :: ctxt -> String -> v -> v

sjoin_pointers :: ctxt -> [p] -> [p]

top :: ctxt -> String -> v

ssemantics :: ctxt -> String -> SymbolicOperation v -> v

sflg_semantics :: ctxt -> v -> Instruction -> FlagStatus -> FlagStatus

simmediate :: Integral i => ctxt -> i -> v

smk_init_reg_value :: ctxt -> Register -> v

smk_init_mem_value :: ctxt -> String -> p -> RegionSize -> v

scall :: ctxt -> Bool -> Instruction -> State (Sstate v p, VCS) ()

sjump :: ctxt -> Instruction -> State (Sstate v p, VCS) ()

stry_jump_targets :: ctxt -> v -> Maybe (Set ResolvedJumpTarget)

stry_immediate :: ctxt -> v -> Maybe Word64

stry_deterministic :: ctxt -> v -> Maybe SimpleExpr

stry_relocation :: ctxt -> p -> RegionSize -> Maybe v

saddress_has_instruction :: ctxt -> Word64 -> Bool

data Sstate v p

Constructors

Sstate

Fields

sregs :: Map Register v

smem :: Map (p, RegionSize) v

sflags :: FlagStatus

execSstate :: State (Sstate v p, VCS) b -> Sstate v p -> Sstate v p

evalSstate :: State (Sstate v p, VCS) a -> Sstate v p -> a

sread_rreg :: SymbolicExecutable ctxt p1 p2 => ctxt -> Register -> (Sstate p1 p3, b) -> p1

Read from a register

sread_reg :: SymbolicExecutable ctxt v p => ctxt -> Register -> State (Sstate v p, VCS) v

swrite_rreg :: SymbolicExecutable ctxt v p => ctxt -> Register -> v -> State (Sstate v p, VCS) ()

Write to a register

soverwrite_reg :: SymbolicExecutable ctxt v p => ctxt -> Bool -> Register -> v -> State (Sstate v p, VCS) ()

swrite_reg :: SymbolicExecutable ctxt v p => ctxt -> Register -> v -> State (Sstate v p, VCS) ()

sread_mem :: SymbolicExecutable ctxt v p => ctxt -> String -> v -> RegionSize -> State (Sstate v p, VCS) v

Read from memory

sread_mem_from_ptr :: SymbolicExecutable ctxt v p => ctxt -> String -> p -> RegionSize -> State (Sstate v p, VCS) v

swrite_mem :: SymbolicExecutable ctxt v p => ctxt -> Bool -> v -> RegionSize -> v -> State (Sstate v p, VCS) ()

Write to memory

swrite_mem_to_ptr :: SymbolicExecutable ctxt v p => ctxt -> Bool -> p -> RegionSize -> v -> State (Sstate v p, VCS) ()

sresolve_address :: SymbolicExecutable ctxt v p => ctxt -> Address -> State (Sstate v p, VCS) v

Given the address of an operand of an instruction, resolve it given the current state.

sread_operand :: SymbolicExecutable ctxt v p => ctxt -> String -> Operand -> State (Sstate v p, VCS) v

swrite_operand :: SymbolicExecutable ctxt v p => ctxt -> Bool -> Operand -> v -> State (Sstate v p, VCS) ()

swrite_flags :: SymbolicExecutable ctxt v p => ctxt -> v -> Instruction -> State (Sstate v p, VCS) ()

clean_flg :: StatePart -> FlagStatus -> FlagStatus

add_jump_to_pred :: Instruction -> Instruction -> FlagStatus -> FlagStatus

sreturn :: SymbolicExecutable ctxt v p => ctxt -> State (Sstate v p, VCS) ()

slea :: SymbolicExecutable ctxt v p => ctxt -> AddressWord64 -> Operand -> Operand -> State (Sstate v p, VCS) ()

smov :: SymbolicExecutable ctxt v p => ctxt -> a -> Instruction -> State (Sstate v p, VCS) ()

sgeneric_cinstr :: SymbolicExecutable ctxt v p => ctxt -> Instruction -> State (Sstate v p, VCS) ()

operand_size :: HasSize a => GenericOperand a -> Int

maybe_operand_size :: HasSize storage => [GenericOperand storage] -> Maybe Int

sexec_cinstr :: SymbolicExecutable ctxt v p => ctxt -> Instruction -> State (Sstate v p, VCS) ()

sset_rip :: SymbolicExecutable ctxt v p => ctxt -> Instruction -> State (Sstate v p, VCS) ()

sexec_instr :: SymbolicExecutable ctxt v p => ctxt -> Instruction -> State (Sstate v p, VCS) ()

sexec_block :: SymbolicExecutable ctxt v p => ctxt -> [GenericInstruction AddressWord64 Register Prefix Opcode Int] -> Maybe [GenericInstruction AddressWord64 Register Prefix Opcode Int] -> Sstate v p -> (Sstate v p, Set VerificationCondition)

sjoin_mem :: SymbolicExecutable ctxt v p => ctxt -> String -> Sstate v p -> Sstate v p -> Map (p, RegionSize) v

sjoin_states :: forall v p1 ctxt p2. (SymbolicExecutable ctxt v p1, SymbolicExecutable ctxt v p2) => ctxt -> [Char] -> Sstate v p1 -> Sstate v p1 -> Sstate v p1

supremum :: SymbolicExecutable ctxt v p => ctxt -> [Sstate v p] -> Sstate v p

The supremum of a list of predicates

simplies :: SymbolicExecutable ctxt v p => ctxt -> Sstate v p -> Sstate v p -> Bool