Instantiation.SymbolicPropagation

Description

Documentation

get_invariant :: FContext -> Int -> Maybe Predicate

Get the currently known invariant for the given instruction address

init_pred

Arguments

:: FContext
The current context

-> String
The current function

-> Invariants
The currently available invariants

-> Set (NodeInfo, Predicate)
The currently known postconditions

-> Set SStatePart
The currently known stateparts of the function

-> Predicate

The initial predicate.

invariant_to_finit :: FContext -> Predicate -> FInit

Convert the current invariant into a function initialisation

join_finit :: FContext -> FInit -> FInit -> FInit

The join between two function initialisations

gather_stateparts

Arguments

:: Invariants
The currently available invariants

-> Set (NodeInfo, Predicate)
The currently known postconditions

-> Set SStatePart

Given the currently known invariants and postconditions, gather all stateparts occurring in the current function.

Orphan instances

Propagator FContext Predicate

Instance details

Methods

tau :: FContext -> [Instruction] -> Maybe [Instruction] -> Predicate -> (Predicate, Set VerificationCondition)

join :: FContext -> String -> Predicate -> Predicate -> Predicate

implies :: FContext -> Predicate -> Predicate -> Bool

SymbolicExecutable FContext SValue SPointer

Instance details

Methods

sseparate :: FContext -> String -> SPointer -> RegionSize -> SPointer -> RegionSize -> Bool

senclosed :: FContext -> SPointer -> RegionSize -> SPointer -> RegionSize -> Bool

salias :: FContext -> SPointer -> RegionSize -> SPointer -> RegionSize -> Bool

ssensitive :: FContext -> SPointer -> RegionSize -> SValue -> Bool

sread_from_ro_data :: FContext -> SPointer -> RegionSize -> Maybe SValue

smk_mem_addresses :: FContext -> String -> SValue -> Set SPointer

sjoin_values :: Foldable t => FContext -> String -> t SValue -> SValue

swiden_values :: FContext -> String -> SValue -> SValue

sjoin_pointers :: FContext -> [SPointer] -> [SPointer]

top :: FContext -> String -> SValue

ssemantics :: FContext -> String -> SymbolicOperation SValue -> SValue

sflg_semantics :: FContext -> SValue -> Instruction -> FlagStatus -> FlagStatus

simmediate :: Integral i => FContext -> i -> SValue

smk_init_reg_value :: FContext -> Register -> SValue

smk_init_mem_value :: FContext -> String -> SPointer -> RegionSize -> SValue

scall :: FContext -> Bool -> Instruction -> State (Sstate SValue SPointer, VCS) ()

sjump :: FContext -> Instruction -> State (Sstate SValue SPointer, VCS) ()

stry_jump_targets :: FContext -> SValue -> Maybe (Set ResolvedJumpTarget)

stry_immediate :: FContext -> SValue -> Maybe Word64

stry_deterministic :: FContext -> SValue -> Maybe SimpleExpr

stry_relocation :: FContext -> SPointer -> RegionSize -> Maybe SValue

saddress_has_instruction :: FContext -> Word64 -> Bool