MachineState

Description

These functions are defined using the State (Pred,VCS) monad. Both the read- and write function may update the current predicate, as well as introduce new verification conditions.

Documentation

read_reg :: Context -> Register -> State (Pred, VCS) SimpleExpr

Read from a register

write_reg :: Context -> Register -> SimpleExpr -> State (Pred, VCS) ()

Write to a register

read_mem

Arguments

:: Context
The context

-> Address
The address of an operand of an instruction

-> State (Pred, VCS) SimpleExpr

Read from memory

write_mem

Arguments

:: Context
The context

-> MemWriteIdentifier
An identifier where the write occurs

-> SimpleExpr
The symbolic address

-> Int
The size (in bytes)

-> SimpleExpr
The value to be written

-> State (Pred, VCS) ()

Write to memory Each memory write is accomponied with a MemWriteIdentifier so that we can log memory writes to unknown locations.

read_operand :: Context -> Operand -> State (Pred, VCS) SimpleExpr

Read from an operand of an instruction

write_operand

Arguments

:: Context
The context

-> Int
The address of the instruction, used to build a MemWriteIdentifier

-> Operand
The operand

-> SimpleExpr
The value to be written

-> State (Pred, VCS) ()

Write to an operand of an instruction

read_sp :: Context -> StatePart -> State (Pred, VCS) SimpleExpr

Read from an statepart

write_sp

Arguments

:: Context
The context

-> (StatePart -> MemWriteIdentifier)

-> (StatePart, SimpleExpr)

-> State (Pred, VCS) ()

Write to a statepart

invalid_bottom_pointer :: Context -> SimpleExpr -> Bool

Returns true if a pointer is not suitable for writing to memory. This may happen if the symbolic expression provides no information, i.e., it has Bottom without known pointerbases, and without any sources.

address_is_unwritable :: Context -> SimpleExpr -> Bool

An address is considered "unwritable" only if it is an immediate address that belongs to a section that is considered unwritable according to Conventions (see section_is_unwritable)

resolve_address :: Context -> Address -> State (Pred, VCS) SimpleExpr

Given the address of an operand of an instruction, resolve it given the current state.