NASM.NASM

Description

Documentation

data NASM

NASM contains external symbols, sections, and a footer

Constructors

NASM

Fields

nasm_externals :: Set String

nasm_sections :: [NASM_Section]

nasm_footer :: [String]

data NASM_Label

A NASM label is either a string, or a macro. The latter is used when referring to addresses within data sections.

Constructors

Label String
Normal label

Macro String String Word64 Word64
Macro into data section (segment, section, a0, offset)

data NASM_Section

A NASM section is either a NASM text section or NASM data section A NASM data section is simply a String

Constructors

NASM_Section_Text NASM_TextSection

NASM_Section_Data [NASM_DataSection]

type Annot = [(Word64, NASM_Label, Word64)]

An annotation consists of an address that is being symbolized to a label and an offset. The offset will often be 0. For example: 0x1016 --> L1000_2 0x4028 --> L_.bss_0x4020 + 0x8

data NASM_TextSection

A NASM text section contains a header (usually just some comments), and a graph of basic blocks

Constructors

NASM_TextSection

Fields

nasm_function_name :: String

nasm_blocks :: [(Int, [NASM_Line])]

A mapping of blockIDs to instructions

data NASM_Line

A line is either a comment, an instruction or a label

Constructors

NASM_Comment Int String
A comment with an indentation level (number of spaces

NASM_Line NASM_Instruction
An instruction with an annotation

NASM_Label NASM_Label
A label