OutputGeneration.NASM.NASM

Description

Documentation

data NASM

NASM contains external symbols, sections, and a footer

Constructors

NASM

Fields

nasm_externals :: Set String

nasm_globals :: Set String

nasm_sections :: [NASM_Section]

nasm_footer :: [String]

data NASM_Label

A NASM label is either a string, or a macro. The latter is used when referring to addresses within data sections. For example:

Macro "" ".data" 0x4000 0x23

refers to segment "", and the section ".data" that starts at immediate address 0x4000. So the real adress was 0x4023.

Constructors

Label Word64 String
Normal label

Macro String String Word64 Word64
Macro into data section (segment, section, a0, offset)

data NASM_Section

A NASM section is either a NASM text section or NASM data section

Constructors

NASM_Section_Text NASM_TextSection

NASM_Section_Data [NASM_DataSection]

type Annot = [(Word64, NASM_Label, Word64)]

An annotation consists of an address that is being symbolized to a label and an offset. The offset will often be 0. For example: 0x1016 --> L1000_2 0x4028 --> L_.bss_0x4020 + 0x8 Annotations translate to NASM comments, and have no effect on the actual NASM itself.

data NASM_TextSection

A NASM text section contains a name and an **ordered** list of basic blocks. Each basic block has an ID and a list of lines.

Constructors

NASM_TextSection

Fields

nasm_function_name :: String

nasm_blocks :: [(Int, [NASM_Line])]

A mapping of blockIDs to instructions

nasm_cfg :: IntMap IntSet

data NASM_Line

A NASM line is either a comment, an instruction or a label

Constructors

NASM_Comment Int String
A comment with an indentation level (number of spaces)

NASM_Line NASM_Instruction
An instruction with an annotation

NASM_Label NASM_Label
A label

data NASM_Instruction

An instruction consists of a prefix, an opcode, a list of operands, a comment (possibly empty) and an annotation (possibly empty).

Constructors

NASM_Instruction

Fields

nasm_prefix :: Maybe Prefix

nasm_mnemonic :: Maybe Opcode

nasm_operands :: [NASM_Operand]

nasm_comment :: String

nasm_annot :: Annot

type NASM_SizeDir = (Int, Bool)

A size directive for a memory operand in bytes. For example (4,True) is the size directive for a 4-byte memory operand. It can be False to indicate that the operand should not be rendered (e.g, in case of an LEA instruction).

data NASM_Address

An address can either be a computation or some symbol.

Constructors

NASM_Addr_Compute NASM_Address_Computation

NASM_Addr_Symbol Symbol

NASM_Addr_Label NASM_Label (Maybe Word64)

NASM_JumpTarget ResolvedJumpTarget

data NASM_Address_Computation

An address computation within an operand. The computation is: segment + [base + index*scale + disp]

Constructors

NASM_Address_Computation

Fields

nasm_segment :: Maybe Register

nasm_index :: Maybe Register

nasm_scale :: Word64

nasm_base :: Maybe Register

nasm_displace :: Maybe Word64

data NASM_DataEntry

A data section consists of a list of data section entries. Each DataEntry stores its address and a value.

Constructors

DataEntry_Byte Word64 Word8
A single byte

DataEntry_String Word64 [Word8]
A string of characters (Word8)

DataEntry_Pointer Word64 (NASM_Address, Annot)
A pointer (a label or external symbol)

DataEntry_BSS Int
A BSS section with a given size in bytes

data NASM_DataSection

A data section then consists of:

Constructors

NASM_DataSection

Fields

nasm_data_section :: (String, String, Word64)

(segment,section,address)

nasm_data_section_align :: Int

The alignment (0 if unknown)

nasm_data_section_labels :: IntMap (Set NASM_Label)

Used internally only

nasm_data_section_data :: [NASM_DataEntry]

A list of DataEntries