OutputGeneration.Retrieval

Documentation

ctxt_get_instructions :: Context -> Set Instruction

Retrieves a set of instructions.

ctxt_get_function_entries :: Context -> Set Int

Retrieves a set of funtion entries.

ctxt_get_function_summary :: Num a => Context -> Int -> (a, (FInit, Maybe FReturnBehavior))

Retrieves function summaries (preconditions, postconditions)

ctxt_mk_function_boundary :: Num a => Context -> Int -> (a, [Char])

Retrieves a textual representation of the function boundaries

ctxt_get_inv :: Context -> Word64 -> (Word64, Maybe [(Word64, Predicate)])

Given an address a, retrieve all invariants (the address may occur in multiple functions).

ctxt_resolve_mem_operands :: Context -> [(Word64, Word64, [Maybe SValue])]

Retrieve an overview of memory operands Returns per instruction a tuple (entry, addr, ptrs). Here: entry is the entry address of the function of the instruction, addr is the address of the instruction ptrs is a list of symbolic pointers for each operand (or Nothing if not a memory-operand)

ctxt_get_instruction :: Context -> Int -> IO Instruction

Retrieve instruction for a given instruction address, both as datastructure and pretty-printed

ctxt_get_instruction_addresses :: Context -> Set Word64

Retrieve all instruction addresses.

ctxt_get_controlflow :: Context -> Word64 -> IO (Word64, [Word64])

Given an address a, retrieve the set of next addresses.