Pointers

Description

Documentation

get_pointer_bases :: Context -> SimpleExpr -> Set PointerBase

A PointerBase is a positive addend of a symbolic expression that may represent a pointer. Retrieves the pointer bases from a symbolic expression. They are either 1.) all known, or 2.) all unknown.

get_known_pointer_bases :: Context -> SimpleExpr -> Set PointerBase

Returns the set of known pointerbases, or the empty set if none.

expr_highly_likely_pointer :: Context -> SimpleExpr -> Bool

Returns true if the expression has known pointerbases.

expr_is_global_pointer :: Context -> SimpleExpr -> Bool

Returns true if the expression has a global pointerbase.

expr_is_highly_likely_local_pointer :: Context -> SimpleExpr -> Bool

Returns true if the expression has a local pointerbase, and no others.

expr_is_possibly_local_pointer :: Context -> SimpleExpr -> Bool

Returns true if the expression has a local pointerbase, but maybe other pointerbases as well.

expr_is_global_immediate :: Context -> SimpleExpr -> Bool

Returns true iff the expression is an immediate address falling into the range of sections of the binary

srcs_of_expr :: Context -> SimpleExpr -> Set BotSrc

Returns the set of sources (inputs used to compute the expression) of an expression.

srcs_of_exprs :: Context -> [SimpleExpr] -> Set BotSrc

Returns the set of sources (state parts used to compute the expression) of two expressions.

pointers_from_different_global_section :: Context -> SimpleExpr -> SimpleExpr -> Bool

Returns true iff the two given expressions have global pointerbases in different segments/sections of the binary. We do not assume that such pointers are separate, but do assert it.

join_exprs :: [Char] -> Context -> [SimpleExpr] -> SimpleExpr

join_single :: Context -> SimpleExpr -> SimpleExpr

Abstraction for a single expression, even if the expression is concrete.

pointer_bases_separate :: PointerBase -> PointerBase -> Bool

Two pointerbases are separate if they refer to completely different parts of the memory. We assume Stackframe, Global address space, and Heap are separate. Two different malloc's point to different regions.

sources_separate :: Context -> FInit -> BotSrc -> BotSrc -> Bool

Two sources are inputs for separate pointers if, e.g., one of them is the stackpointer and the other a malloc-return-value.

necessarily_equal :: SimpleExpr -> SimpleExpr -> Bool

Returns true iff the given symbolic regions are necessarily equal. For example: [RSP-16,4] is enclosed in [RSP-16,4]

Will return False if the expressions contain bottom.

necessarily_equal_stateparts :: StatePart -> StatePart -> Bool

Returns true iff the given symbolic stateparts are necessarily equal.

necessarily_separate :: (Ord a, Num a) => Context -> FInit -> SimpleExpr -> a -> SimpleExpr -> a -> Bool

Returns true iff the given symbolic regions are ncessarily separate. For example: [RSP-16,4] is separate from [RSP-12,8] [RSP+8,4] is separate from [RSP,8]

If none of the cases apply where it can be shjown arithmetically that the expressions are separate, we check whether the expressions can be proven separate based on their domains (see separate_pointer_domains).

necessarily_separate_stateparts :: Context -> FInit -> StatePart -> StatePart -> Bool

Returns true iff the given symbolic stateparts are necessarily separate.

necessarily_enclosed :: Integral a => SimpleExpr -> a -> SimpleExpr -> a -> Bool

Returns true iff the given symbolic region is necessarily enclosed in the other. For example: [RSP-16,4] is enclosed in [RSP-18,8] [RSP+4,4] is enclosed in [RSP,8]

Will return False if the expressions contain bottom.