Propagation

Description

We assume a class where we can do predicate transformation through function tau, and we can merge two predicates through function join. Moreover, we assume an implementation of a function implies that implements symbolic implication.

Documentation

class Show pred => Propagator ctxt pred where

A class that allows propagation of predicates over a CFG.

Methods

tau :: ctxt -> CFG -> Int -> Maybe Int -> pred -> pred

Predicate transformation for an edge in in a CFG, over a basic blocks.

join :: ctxt -> pred -> pred -> pred

A lattice-join

implies :: ctxt -> pred -> pred -> Bool

Symbolic implication

do_prop

Arguments

:: Propagator ctxt pred

=> ctxt
The context

-> CFG
The CFG

-> Int
The entry address

-> pred
The initial predicate

-> IntMap pred

Start propagation at the given entry address with the given initial predicate. Returns a set of invariants, i.e., a mapping of instruction addresses to predicates.

post :: CFG -> Key -> IntSet

The set of next blocks from the given block