SymbolicExecution

Documentation

tau_blockID

Arguments

:: Context
The context

-> CFG
The CFG

-> Int
The block ID of the basic block

-> Maybe Int
Optionally, the block ID of the next block if symbolically executing an edge in a CFG.

-> Bool
Must the last instruction of the block be ignored?

-> Pred
The predicate to be transformed

-> Pred

Do predicate transformation over a basic block in a CFG. Given an edge in the CFG from blockId to blockId', perform predicate transformation. blockId' is needed to set the flags properly. If none is supplied, the flags are ignored. We provide the option to skip the last instruction, as sometimes we need to symbolically execute a block up to but excluding the last instruction.

tau_b :: Context -> [Instr] -> State Pred ()

Do predicate transformation over a block of instructions Does not take into account flags, commonly function tau_blockID should be used.

init_pred :: Invariants -> Set (NodeInfo, Pred) -> Pred

The initial predicate Given the currently known invariants and postconditions, gather all stateparts occurring in the current function. The initial predicate assign a variable to each statepart, e.g.: RSP == RSP0, RDI == RDI), ....

statepart_is_preserved_after_function_call

Arguments

:: Context
The context

-> Instr
The instruction currently symbolically executed (a call)

-> StatePart
A state part

-> Bool

Returns true iff the given statepart is preserved (unmodified) after a function call.

gather_stateparts :: Invariants -> Set (NodeInfo, Pred) -> Set StatePart

expr_highly_likely_pointer :: Context -> SimpleExpr -> Bool

Tries to guess if the symbolic expression is a pointer. Any immediate that falls in the address range of the binary is highly likely a pointer. A value stored in a memory location belonging to an external function symbol is also highly likely a pointer.

Orphan instances

Propagator Context Pred

Instance details

Methods

tau :: Context -> CFG -> Int -> Maybe Int -> Pred -> Pred

join :: Context -> Pred -> Pred -> Pred

implies :: Context -> Pred -> Pred -> Bool