WithAbstractSymbolicValues.Class

Documentation

data SymbolicOperation v

Constructors

SO_Op Opcode Int (Maybe Int) [v]

SO_Bit Int v

SO_SExtend Int Int v

SO_Overwrite Int v v

SO_Plus v v

SO_Minus v v

SO_Times v v

data SStatePart p

A statepart is either a register or a region in memory

Constructors

SSP_Reg Register
A register

SSP_Mem p Int
A region with a symbolic address and an immediate size.

data Sstate v p

Constructors

Sstate

Fields

sregs :: Map Register v

smem :: Map (p, Maybe ByteSize) v

sflags :: FlagStatus

data MemRelation

A function initialisation consists of a mapping of state parts to values, and memory relations

Constructors

Separate

Aliassing

Unknown

data FInit v p

Constructors

FInit (Set (SStatePart p, v)) (Map (SStatePart p, SStatePart p) MemRelation)

empty_finit :: FInit v p

unknownSize :: Maybe a

class (Ord v, Eq v, Show v, Eq p, Ord p, Show p) => WithAbstractSymbolicValues ctxt v p | ctxt -> v p where

Methods

sseparate :: ctxt -> String -> p -> Maybe ByteSize -> p -> Maybe ByteSize -> Bool

senclosed :: ctxt -> p -> Maybe ByteSize -> p -> Maybe ByteSize -> Bool

salias :: ctxt -> p -> Maybe ByteSize -> p -> Maybe ByteSize -> Bool

ssensitive :: ctxt -> p -> Maybe ByteSize -> v -> Bool

sread_from_ro_data :: ctxt -> p -> Maybe ByteSize -> Maybe v

smk_mem_addresses :: ctxt -> String -> v -> Set p

sjoin_values :: Foldable t => ctxt -> String -> t v -> v

swiden_values :: ctxt -> String -> v -> v

sjoin_pointers :: ctxt -> [p] -> [p]

top :: ctxt -> String -> v

ssemantics :: ctxt -> String -> SymbolicOperation v -> v

sflg_semantics :: ctxt -> v -> Instruction -> FlagStatus -> FlagStatus

simmediate :: Integral i => ctxt -> i -> v

smk_init_reg_value :: ctxt -> Register -> v

smk_init_mem_value :: ctxt -> String -> p -> Maybe ByteSize -> v

scall :: ctxt -> Instruction -> State (Sstate v p, VCS v) ()

sjump :: ctxt -> Instruction -> State (Sstate v p, VCS v) ()

saddress_has_instruction :: ctxt -> Word64 -> Bool

skeep_for_finit :: ctxt -> SStatePart p -> v -> Maybe (Set p)

stry_jump_targets :: ctxt -> v -> Maybe (Set ResolvedJumpTarget)

stry_resolve_error_call :: ctxt -> Instruction -> v -> Maybe Indirection

stry_immediate :: ctxt -> v -> Maybe Word64

sis_deterministic :: ctxt -> v -> Bool

scheck_regs_in_postcondition :: ctxt -> v -> v -> Bool