X86.Instruction

Documentation

type Instruction = GenericInstruction AddressWord64 Register Prefix Opcode Int

canonicalize :: Instruction -> [Instruction]

Canonicalizes an instruction by making sure there is at most one output and inputs/outputs are syntactically clear. | For example, the IMUL instruction comes in different flavors that hide the actual inputs and outputs to multiple registers

Orphan instances

HasAddress Instruction

Instance details

Methods

addressof :: Instruction -> Word64

HasSize Instruction

Instance details

Methods

sizeof :: Instruction -> Int