Analysis.Context

Description

The context stores, among others, information obtained during verification, such as CFGs, invariants, etc. (see Context). Module VerificationReportInterface provides functions for obtaining and interfacing with a Context.

Documentation

read_binary :: String -> IO (Maybe Binary)

data CFG

A control flow graph with blocks and edges. A blockID (represented as an Int) is a unique identifier of a basic block. We store basic blocks twice: once as addresses, and once as instructions.

Constructors

CFG

Fields

cfg_blocks :: IntMap [Int]

A mapping of blockIDs to instruction addresses

cfg_edges :: IntMap IntSet

A mapping of blockIDs to sets of blocKIDs

cfg_addr_to_blockID :: IntMap Int

A mapping of instruction addresses to blockIDs

cfg_fresh :: Int

A fresh blockID

cfg_instrs :: IntMap [Instruction]

A mapping of blockIDs to lists of disassembled instructions.

data JumpTable

A jump table Let the actual indirection be JMP RAX, implying jt_trgt_operand == RAX. This jump table can be implemented with inserting this instruction before the indirection:

MOV jt_trgt_operand, QWORD PTR [jt_address + 8*jt_index_operand]

Constructors

JumpTable

Fields

jt_index_operand :: Operand

The operand that is bounded by some immediate, serving as an index into a table

jt_trgt_operand :: Operand

The operand of the jump

jt_table_entries :: [Int]

An ordered list of instruction addresses to which is jumped

data ResolvedJumpTarget

Resolving the operand of a jump/call can produce one of the following.

Constructors

Unresolved
An indirect branch that has not been resolved yet

External String
A call to external function f

ImmediateAddress Word64
An internal call to the given address

data Indirection

An indirection

Constructors

IndirectionResolved (Set ResolvedJumpTarget)
An indirection that could be resolved to one or more addresses

IndirectionJumpTable JumpTable
An indirection based on a jump table

type Indirections = IntMap Indirection

Per instruction address, a a jump table

data VerificationResult

An enumeration indicating the result of verification over a function

Constructors

VerificationSuccess
Function was succesfully verified

VerificationSuccesWithAssumptions
Function was succesfully verified, but required assertions

VerificationUnresolvedIndirection
Function contains an unresolved indirection

VerificationError String
There was some verification error, e.g., return adresss overwrite

Unverified
The function has not been verified.

type Invariants = IntMap Pred

Invariants: a mapping of blockIDs to predicates

data NodeInfo

For each leaf-node in a CFG we store the following info.

Constructors

Normal
The basic block behaves normally, e.g., a ret

UnresolvedIndirection
The basic block ends in an unresolved indirection

Terminal
The basic blocks ends with, e.g., a call to exit()

type Postconditions = Set (NodeInfo, Pred)

Postconditions: for each final block the NodeInfo and the final predicate after execution of the block

data MemWriteIdentifier

Identifies where a memwrite occurred TODO should be generalized, is a StatePartWriteIdentifier

Constructors

MemWriteFunction String Word64 StatePart
A function with name at address i_a wrote to a statepart

MemWriteInstruction Word64 Operand SimpleExpr
An instruction wrote to an operand, resolving to an address

data VerificationCondition

A verification condition is either: * A precondition of the form: > Precondition (a0,si0) (a1,si1) This formulates that at the initial state the two regions must be separate. * An assertion of the form: > Assertion a (a0,si0) (a1,si1) This formulates that dynamically, whenever address a is executed, the two regions are asserted to be separate. * A function constraint of the form: > FunctionConstraint foo [(RDI, v0), (RSI, v1), ...] { sp0,sp1,... } This formulates that a function call to function foo with values v0, v1, ... stored in the registers should not overwrite certain state parts.

Constructors

Precondition SimpleExpr Int SimpleExpr Int
Precondition: lhs SEP rhs

Assertion SimpleExpr SimpleExpr Int SimpleExpr Int
Assertion: @address, lhs SEP rhs

FunctionConstraint String Word64 [(Register, SimpleExpr)] (Set StatePart)
Function name, address, of call, with param registers

SourcelessMemWrite MemWriteIdentifier
A write to a statepart for which no information was available

FunctionPointers Word64 IntSet
A set of function pointers passed to a function

type VCS = Set VerificationCondition

An acornym for a set of verification conditions

data PointerDomain

An abstract domain for pointers

Constructors

Domain_Bases (Set PointerBase)

Domain_Sources (Set BotSrc)

type FInit = Map StatePart SimpleExpr

A function initialisation consists of a mapping of stateparts to expressions.

data FReturnBehavior

A function call

Constructors

Terminating
The function does never return

ReturningWith Pred
The function returns withg the symbolic changes stored in the predicate

UnknownRetBehavior
It is unknown whether the function returns or not

data Context_

The context datastructure.

S: Information Statically obtained by reading from the binary

D: Information Dynamically updated during verification

Constructors

Context_

Fields

ctxt__config :: !Config

S: the configuration file in ./config ctxtdata :: !(IM.IntMap Word8), -- ^ S__: mapping from addresses to bytes (writable data section)

ctxt__syms :: !(IntMap String)

S: the symbol table: a mapping of addresses to function names for external functions

ctxt__sections :: !SectionsInfo

S: information on segments/section

ctxt__dirname :: !String

S: the name of the directory where the .dump, .entry, .sections and .symbols files reside

ctxt__name :: !String

S: the name of the binary

ctxt__start :: !Int

S: the address of the _start symbol

ctxt__relocs :: ![Relocation]

S: relocation entries

ctxt__entries :: !Graph

D: a graph with an edge (e0,e1) if entry address e0 calls entry address e1, and e0 and e1 have not been verified yet

ctxt__cfgs :: !(IntMap CFG)

D: the currently known control flow graphs per function entry

ctxt__calls :: !(IntMap FReturnBehavior)

D: the currently known and verified entry addresses of functions mapped to return-information

ctxt__invs :: !(IntMap Invariants)

D: the currently known invariants

ctxt__posts :: !(IntMap Postconditions)

D: the currently known postconditions

ctxt__inds :: !Indirections

D: the currently known indirections

ctxt__finits :: !(IntMap FInit)

D: the currently known function initialisations

ctxt__vcs :: !(IntMap VCS)

D: the verification conditions

ctxt__results :: !(IntMap VerificationResult)

D: the verification result

ctxt__recursions :: !(IntMap IntSet)

D: a mapping from function entries to the set of mutually recursive functions entries they occur in

data Context

Constructors

Context

Fields

ctxt_binary :: !Binary

ctxt_ioref :: IORef (IntMap Instruction)

ctxt_ctxt_ :: !Context_

ctxt_config :: Context -> Config

ctxt_syms :: Context -> IntMap String

ctxt_sections :: Context -> SectionsInfo

ctxt_dirname :: Context -> String

ctxt_name :: Context -> String

ctxt_start :: Context -> Int

ctxt_relocs :: Context -> [Relocation]

ctxt_entries :: Context -> Graph

ctxt_cfgs :: Context -> IntMap CFG

ctxt_calls :: Context -> IntMap FReturnBehavior

ctxt_invs :: Context -> IntMap Invariants

ctxt_posts :: Context -> IntMap Postconditions

ctxt_inds :: Context -> Indirections

ctxt_finits :: Context -> IntMap FInit

ctxt_vcs :: Context -> IntMap VCS

ctxt_results :: Context -> IntMap VerificationResult

ctxt_recursions :: Context -> IntMap IntSet

set_ctxt_syms :: IntMap String -> Context -> Context

set_ctxt_sections :: SectionsInfo -> Context -> Context

set_ctxt_start :: Int -> Context -> Context

set_ctxt_entries :: Graph -> Context -> Context

set_ctxt_calls :: IntMap FReturnBehavior -> Context -> Context

set_ctxt_inds :: Indirections -> Context -> Context

set_ctxt_posts :: IntMap Postconditions -> Context -> Context

set_ctxt_vcs :: IntMap VCS -> Context -> Context

set_ctxt_finits :: IntMap FInit -> Context -> Context

set_ctxt_invs :: IntMap Invariants -> Context -> Context

set_ctxt_cfgs :: IntMap CFG -> Context -> Context

set_ctxt_relocs :: [Relocation] -> Context -> Context

set_ctxt_results :: IntMap VerificationResult -> Context -> Context

set_ctxt_recursions :: IntMap IntSet -> Context -> Context

init_context :: Binary -> IORef (IntMap Instruction) -> Config -> String -> String -> Context

intialize an empty context based on the command-line parameters

purge_context :: Context -> Context_

purge the context before exporting it (may save a lot of disk space)

read_from_ro_datasection

Arguments

:: Context
The context

-> Word64
An address

-> Int
Size, i.e., the number of bytes to read

-> Maybe Word64

Reading from a read-only data section.

Reads maximally up to 8 bytes. Returns Nothing if the given address is out-of-range.

is_roughly_an_address

Arguments

:: Context
The context

-> Word64
An address

-> Bool

Is the immediate roughly in range to be an address?

find_section_for_address

Arguments

:: Context
The context

-> Word64
An address

-> Maybe (String, String, Word64, Word64)

Find a section for an address (see SectionsInfo)

fetch_instruction

Arguments

:: Context
The context

-> Word64
An address

-> IO (Maybe Instruction)

Fetching an instruction

Returns Nothing if the given address is out-of-range.

pp_instruction

Arguments

:: Context
The context

-> Instruction
An instruction

-> String

Pretty printing an instruction

show_finit :: FInit -> String

Show function initialisaExpr

is_assertion :: VerificationCondition -> Bool

Is the given verification condition an assertion?

is_precondition :: VerificationCondition -> Bool

Is the given verification condition a precondition?

is_func_constraint :: VerificationCondition -> Bool

Is the given verification condition a function constraint?

is_sourceless_memwrite :: VerificationCondition -> Bool

Is the given verification condition a sourceless memwrite?

is_functionpointers :: VerificationCondition -> Bool

Is the given verification condition a function pointer introduction?

count_instructions_with_assertions :: Set VerificationCondition -> Int

Count the number of assertions in the set of verification conditions.

count_sourceless_memwrites :: Set VerificationCondition -> Int

Count the number of sourceless memory writes in the set of verification conditions.

ctxt_continue_on_unknown_instruction :: Context -> Bool

ctxt_generate_pdfs :: Context -> Bool

ctxt_verbose_logs :: Context -> Bool

ctxt_store_preconditions_in_report :: Context -> Bool

ctxt_store_assertions_in_report :: Context -> Bool

ctxt_max_time :: Context -> Int

ctxt_max_num_of_cases :: Context -> Int

ctxt_max_num_of_bases :: Context -> Int

ctxt_max_num_of_sources :: Context -> Int

ctxt_max_jump_table_size :: Context -> Int

ctxt_max_expr_size :: Context -> Int

Orphan instances

Serialize Relocation

Instance details

Methods

put :: Putter Relocation

get :: Get Relocation

Serialize SectionsInfo

Instance details

Methods

put :: Putter SectionsInfo

get :: Get SectionsInfo