Context

Description

The context stores, among others, information obtained during verification, such as CFGs, invariants, etc. (see Context). Module VerificationReportInterface provides functions for obtaining and interfacing with a Context.

Documentation

data CFG

A control flow graph with blocks and edges. A blockID (represented as an Int) is a unique identifier of a basic block. We store basic blocks twice: once as addresses, and once as instructions.

Constructors

CFG

Fields

cfg_blocks :: IntMap [Int]

A mapping of blockIDs to instruction addresses

cfg_edges :: IntMap IntSet

A mapping of blockIDs to sets of blocKIDs

cfg_addr_to_blockID :: IntMap Int

A mapping of instruction addresses to blockIDs

cfg_fresh :: Int

A fresh blockID

cfg_instrs :: IntMap [Instr]

A mapping of blockIDs to lists of disassembled instructions.

data JumpTable

A jump table Let the actual indirection be JMP RAX, implying jt_trgt_operand == RAX. This jump table can be implemented with inserting this instruction before the indirection:

MOV jt_trgt_operand, QWORD PTR [jt_address + 8*jt_index_operand]

Constructors

JumpTable

Fields

jt_index_operand :: Operand

The operand that is bounded by some immediate, serving as an index into a table

jt_trgt_operand :: Operand

The operand of the jump

jt_table_entries :: [Int]

An ordered list of instruction addresses to which is jumped

data ResolvedJumpTarget

Resolving the operand of a jump/call can produce one of the following.

Constructors

Unresolved
An indirect branch that has not been resolved yet

External String
A call to external function f

ImmediateAddress Word64
An internal call to the given address

data Indirection

An indirection

Constructors

IndirectionResolved (Set ResolvedJumpTarget)
An indirection that could be resolved to one or more addresses

IndirectionJumpTable JumpTable
An indirection based on a jump table

type Indirections = IntMap Indirection

Per instruction address, a a jump table

type SectionsInfo = [(String, String, Int, Int)]

Sections: segment names, section names, addresses and sizes.

data VerificationResult

An enumeration indicating the result of verification over a function

Constructors

VerificationSuccess
Function was succesfully verified

VerificationSuccesWithAssumptions
Function was succesfully verified, but required assertions

VerificationUnresolvedIndirection
Function contains an unresolved indirection

VerificationError String
There was some verification error, e.g., return adresss overwrite

Unverified
The function has not been verified.

type Invariants = IntMap Pred

Invariants: a mapping of blockIDs to predicates

data NodeInfo

For each leaf-node in a CFG we store the following info.

Constructors

Normal
The basic block behaves normally, e.g., a ret

UnresolvedIndirection
The basic block ends in an unresolved indirection

Terminal
The basic blocks ends with, e.g., a call to exit()

type Postconditions = Set (NodeInfo, Pred)

Postconditions: for each final block the NodeInfo and the final predicate after execution of the block

data MemWriteIdentifier

Identifies where a memwrite occurred TODO should be generalized, is a StatePartWriteIdentifier

Constructors

MemWriteFunction String Int StatePart
A function with name at address i_a wrote to a statepart

MemWriteInstruction Int Operand SimpleExpr
An instruction wrote to an operand, resolving to an address

data VerificationCondition

A verification condition is either: * A precondition of the form: > Precondition (a0,si0) (a1,si1) This formulates that at the initial state the two regions must be separate. * An assertion of the form: > Assertion a (a0,si0) (a1,si1) This formulates that dynamically, whenever address a is executed, the two regions are asserted to be separate. * A function constraint of the form: > FunctionConstraint foo [(RDI, v0), (RSI, v1), ...] { sp0,sp1,... } This formulates that a function call to function foo with values v0, v1, ... stored in the registers should not overwrite certain state parts.

Constructors

Precondition SimpleExpr Int SimpleExpr Int
Precondition: lhs SEP rhs

Assertion SimpleExpr SimpleExpr Int SimpleExpr Int
Assertion: @address, lhs SEP rhs

FunctionConstraint String Int [(Register, SimpleExpr)] (Set StatePart)
Function name, address, of call, with param registers

SourcelessMemWrite MemWriteIdentifier
A write to a statepart for which no information was available

FunctionPointers Int IntSet
A set of function pointers passed to a function

type VCS = Set VerificationCondition

An acornym for a set of verification conditions

data PointerDomain

An abstract domain for pointers

Constructors

Domain_Bases (Set PointerBase)

Domain_Sources (Set BotSrc)

type FInit = Map StatePart SimpleExpr

A function initialisation consists of a mapping of stateparts to expressions.

data FReturnBehavior

A function call

Constructors

Terminating
The function does never return

ReturningWith Pred
The function returns withg the symbolic changes stored in the predicate

UnknownRetBehavior
It is unknown whether the function returns or not

data Context

The context datastructure.

S: Information Statically obtained by reading from the binary

D: Information Dynamically updated during verification

Constructors

Context

Fields

ctxt_dump :: IntMap Word8

S: mapping from addresses to bytes (constant data and instructions from the binary/executable)

ctxt_data :: IntMap Word8

S: mapping from addresses to bytes (writable data section)

ctxt_syms :: IntMap String

S: the symbol table: a mapping of addresses to function names for external functions

ctxt_sections :: SectionsInfo

S: information on segments/section

ctxt_dirname :: String

S: the name of the directory where the .dump, .entry, .sections and .symbols files reside

ctxt_name :: String

S: the name of the binary

ctxt_generate_pdfs :: Bool

S: do we call graphviz to generate PDFs from .dot files?

ctxt_entries :: Graph

D: a graph with an edge (e0,e1) if entry address e0 calls entry address e1, and e0 and e1 have not been verified yet

ctxt_cfgs :: IntMap CFG

D: the currently known control flow graphs per function entry

ctxt_calls :: IntMap FReturnBehavior

D: the currently known and verified entry addresses of functions mapped to return-information

ctxt_invs :: IntMap Invariants

D: the currently known invariants

ctxt_posts :: IntMap Postconditions

D: the currently known postconditions

ctxt_inds :: Indirections

D: the currently known indirections

ctxt_finits :: IntMap FInit

D: the currently known function initialisations

ctxt_vcs :: IntMap VCS

D: the verification conditions

ctxt_results :: IntMap VerificationResult

D: the verification result

ctxt_recursions :: IntMap IntSet

init_context :: String -> String -> Bool -> Context

read_from_datasection

Arguments

:: Context
The context

-> Word64
An address

-> Int
Size, i.e., the number of bytes to read

-> Maybe Word64

Reading from a data section.

Reads maximally up to 8 bytes. Returns Nothing if the given address is out-of-range.

find_section_for_address

Arguments

:: Context
The context

-> Int
An address

-> Maybe (String, String, Int, Int)

Find a section for an address (see SectionsInfo)

fetch_instruction

Arguments

:: Context
The context

-> Int
An address

-> IO (Maybe Instr)

Fetching an instruction

Returns Nothing if the given address is out-of-range.

pp_instruction

Arguments

:: Context
The context

-> Instr
An instruction

-> String

Pretty printing an instruction

show_finit :: FInit -> String

Show function initialisaExpr

is_assertion :: VerificationCondition -> Bool

Is the given verification condition an assertion?

is_precondition :: VerificationCondition -> Bool

Is the given verification condition a precondition?

is_func_constraint :: VerificationCondition -> Bool

Is the given verification condition a function constraint?

is_sourceless_memwrite :: VerificationCondition -> Bool

Is the given verification condition a sourceless memwrite?

is_functionpointers :: VerificationCondition -> Bool

Is the given verification condition a function pointer introduction?

count_instructions_with_assertions :: Set VerificationCondition -> Int

Count the number of assertions in the set of verification conditions.

count_sourceless_memwrites :: Set VerificationCondition -> Int

Count the number of sourceless memory writes in the set of verification conditions.