Conventions

Description

Documentation

exiting_function_calls :: [[Char]]

A list of function names of functions that never return.

sections_with_instructions :: [([Char], [Char])]

Overview of sections with instructions.

section_is_unwritable :: ([Char], [Char]) -> Bool

Sections in the following list are assumed not to be modifiable during execution, i.e., constant.

callee_saved_registers :: [Register]

A list if registers that are non-volatile, i.e., that must be preserved by a function (callee-saved)

return_registers :: [Register]

A list of registers that may be used for return values

parameter_registers :: [Register]

A list of registers used as parameters