Data.SymbolicExpression

Description

A datatype for symbolic predicates, tailored to storing information on equalities between the current values stored in state parts (lhs) and constant expressions (rhs).

Documentation

data StatePart

A statepart is either a register or a region in memory

Constructors

SP_StackPointer String
The stack pointer of the given function

SP_Reg Register
A register

SP_Mem SimpleExpr Int
A region with a symbolic address and an immediate size.

data SimpleExpr

A symbolic expression with as leafs either immediates, variables, live values of stateparts, or malloced addresses. A variable is a constant representing some initial value, e.g., RDI_0, or [RSP_0,8]_0. A statepart evaluates to its current value, e.g., RDI or [RSP,8].

Constructors

SE_Immediate Word64
An immediate word

SE_Var StatePart
A variable representing the initial value stored in the statepart (e.g., RSP0)

SE_StatePart StatePart
The value stored currently in the statepart

SE_Malloc (Maybe Word64) (Maybe String)
A malloc return value with possibly an ID

SE_Op Operator Int [SimpleExpr]
Application of an Operator to the list of arguments

SE_Bit Int SimpleExpr
Taking the lower bits of a value

SE_SExtend Int Int SimpleExpr
Sign extension

SE_Overwrite Int SimpleExpr SimpleExpr
Overwriting certain bits of a value with bits from another value

Bottom BotTyp
Bottom (unknown value)

data FlagStatus

Symbolically represent the status of all flags in the current state

Constructors

None
No information known, flags could have any value

FS_CMP (Maybe Bool) Operand Operand
The flags are set by the x86 CMP instruction applied to the given operands.

data BotTyp

Bot represents an unknown (bottom) value. We annotate each occurence of Bot with a BotTyp. This type indicates where the bottom value originates from. The latter six are all equal, we just use them for debugging and information. They indicate that the value is unknown, but was computed using the set of sources.

Constructors

FromNonDeterminism (NESet SimpleExpr)
The expression evaluates to one of the expressions in the set

FromPointerBases (NESet PointerBase)
The expression is a pointer-computation with known base(s)

FromCall String
Return value of a function call

FromSources (NESet BotSrc)
The expression is some computation based on sources.

FromOverlap (NESet BotSrc)
A read from two possibly overlapping regions

FromMemWrite (NESet BotSrc)
A write to two possibly overlapping regions

FromSemantics (NESet BotSrc)
An instruction with unknown semantics

FromBitMode (NESet BotSrc)
Should not happen, but if a register writes to a registeralias with unknown bit size

FromUninitializedMemory (NESet BotSrc)
Reading from memory not written to yet

data BotSrc

Sources that may be used to compute an expression. That is, the inputs to an expression.

Constructors

Src_Var StatePart
An initial variable, i.e., a constant

Src_StackPointer String
The stack pointer of the given function

Src_Malloc (Maybe Word64) (Maybe String)
A malloced address

Src_ImmediateAddress Word64
An immediate used in the computation of the pointer

Src_Function String
A return value from a function

Src_Mem (NESet BotSrc)
reading from memory

Src_ImmediateConstants
Some immediate value

data Operator

An operator is a pure operation over bit-vectors, annotated with the bit-size of its operands. For example, Plus 64 denotes 64-bit addition. Udiv and Times are operators op type w -> w -> w with all words same length. Div and Div_Rem are operators of type w -> w -> w -> w performing concatenation of the first two words and then doing division/remainder.

Constructors

Minus

Plus

Times

And

Or

Xor

Not

Bsr

Div_Rem

Div

Shl

Shr

Sar

Udiv

Ror

Rol

Bswap

Pextr

data PointerBase

A pointerbase is a positive addend of a symbolic expression that may represent a pointer.

Constructors

StackPointer String
The stack frame of the given function

Malloc (Maybe Word64) (Maybe String)
A malloc (at the heap) at a given address (hash is unused for now)

GlobalAddress Word64
A global address in the range of the sections of the binary.

PointerToSymbol Word64 String
An address with an associated symbol.

ThreadLocalStorage
A pointer to the thread-local-storage (e.g., FS register)

is_immediate :: SimpleExpr -> Bool

Returns true iff the expression is an immediate value

is_mem_sp :: StatePart -> Bool

Is the statepart memory?

is_reg_sp :: StatePart -> Bool

Is the statepart a register?

contains_bot :: SimpleExpr -> Bool

Returns true iff the expression contains Bot

contains_bot_sp :: StatePart -> Bool

Returns true iff the statepart contains Bot

simp :: SimpleExpr -> SimpleExpr

Simplification of symbolic expressions.

Must always produce an expression logically equivalent to the original. TODO: overwrite(8,Bot[src|RSP_0x1000070ec,malloc100007204()|],b8(Bot[m|RSP_0x1000070ec,malloc100007204()|]))

pp_expr :: SimpleExpr -> [Char]

Pretty print expression, showing Bottom expressions only as Bot

expr_size :: SimpleExpr -> Int

sextend_32_64 :: (Bits p, Num p) => p -> p

sextend_16_64 :: (Bits p, Num p) => p -> p

sextend_8_64 :: (Bits p, Num p) => p -> p