Instantiation.SymbolicPropagation

Description

Documentation

debug_abstraction :: Bool

trace' :: FContext -> String -> a -> a

cspointer_to_exprs :: SPointer -> Set SimpleExpr

ctry_immediate :: SPointer -> Maybe Word64

ctry_deterministic :: SPointer -> Maybe SimpleExpr

cimmediate :: Integral i => i -> SPointer

cwiden :: FContext -> String -> SPointer -> SPointer

cwiden' :: FContext -> SPointer -> SPointer

cwiden_all_to_sources :: FContext -> String -> Set SPointer -> SPointer

cjoin :: FContext -> String -> SPointer -> SPointer -> SPointer

cjoin_all :: Foldable t => FContext -> String -> t SPointer -> SPointer

spointer_plus :: FContext -> Int -> SPointer -> SPointer -> SPointer

spointer_minus :: FContext -> Int -> SPointer -> SPointer -> SPointer

apply_expr_op :: FContext -> ([SimpleExpr] -> Maybe SimpleExpr) -> [SPointer] -> SPointer

data CSemantics

Constructors

ApplyPlus Int

ApplyMinus Int

ApplyNeg Int

ApplyDec Int

ApplyInc Int

Apply ([SimpleExpr] -> SimpleExpr)

SetXX

SExtension_HI

NoSemantics

mk_expr :: FContext -> SimpleExpr -> Maybe SimpleExpr

csemantics :: FContext -> SymbolicOperation SPointer -> SPointer

mnemonic_to_semantics :: Opcode -> Int -> Maybe Int -> CSemantics

cflg_semantics :: p1 -> p2 -> GenericInstruction label Register prefix Opcode annotation -> FlagStatus -> FlagStatus

mk_cvalue :: p -> SimpleExpr -> SPointer

mk_sources :: NESet BotSrc -> SPointer

count_sources :: NESet BotSrc -> Int

mk_cmem_value :: FContext -> SPointer -> p -> SPointer

cmk_mem_addresses :: p -> SPointer -> Set SPointer

srcs_of_spointer :: FContext -> SPointer -> Maybe (NESet BotSrc)

cseparate :: FContext -> [Char] -> SPointer -> SPointer -> SPointer -> SPointer -> Bool

calias :: p -> SPointer -> SPointer -> SPointer -> SPointer -> Bool

cenclosed :: p -> SPointer -> SPointer -> SPointer -> SPointer -> Bool

csensitive :: FContext -> SPointer -> SPointer -> SPointer -> Bool

cread_from_ro_data :: FContext -> SPointer -> SPointer -> Maybe SPointer

get_invariant :: FContext -> Int -> Maybe Predicate

init_pred

Arguments

:: FContext
The current context

-> String
The current function

-> Invariants
The currently available invariants

-> Set (NodeInfo, Predicate)
The currently known postconditions

-> Set SStatePart
The currently known stateparts of the function

-> Predicate

The initial predicate.

gather_stateparts

Arguments

:: Invariants
The currently available invariants

-> Set (NodeInfo, Predicate)
The currently known postconditions

-> Set SStatePart

Given the currently known invariants and postconditions, gather all stateparts occurring in the current function.

invariant_to_finit :: FContext -> Predicate -> FInit

Convert the current invariant into a function initialisation

join_finit :: FContext -> FInit -> FInit -> FInit

The join between two function initialisations

data ExternalFunctionOutput

Constructors

FreshPointer

UnknownReturnValue

Input Register

data ExternalFunctionBehavior

Constructors

ExternalFunctionBehavior

Fields

f_inputs :: [Register]

f_output :: ExternalFunctionOutput

param :: (Eq a, Num a) => a -> Register

pure_and_fresh :: ExternalFunctionBehavior

pure_and_unknown :: ExternalFunctionBehavior

external_function_behavior :: FContext -> String -> ExternalFunctionBehavior

a list of some function that return a heap-pointer through RAX. The pointer is assumed to be fresh.

transpose_bw_exprs :: FContext -> String -> String -> Sstate SPointer -> NESet SimpleExpr -> SPointer

Backward transposition Let p be the current predicate and let the equality sp == v be from the predicate after execution of an internal function. For example, p contains: RSP == RSP0 - 64 RSI == 10

And after execution of the function, we have: *[RSP0+16,8] == RSI0

Transposing this equality produces: *[RSP0-40,8] == 10

transpose_bw_spointer :: FContext -> String -> String -> Sstate SPointer -> SPointer -> SPointer

transpose_bw_reg :: FContext -> String -> String -> Sstate SPointer -> (Register, SPointer) -> Maybe (Register, SPointer)

transpose_bw_mem :: FContext -> String -> String -> Sstate SPointer -> ((SPointer, SPointer), SPointer) -> Maybe ((SPointer, SPointer), SPointer)

transpose_bw_e :: FContext -> String -> String -> Sstate SPointer -> SimpleExpr -> SPointer

transpose_bw_sp :: FContext -> String -> String -> Sstate SPointer -> StatePart -> SStatePart

transpose_bw_bottyp :: FContext -> String -> String -> Sstate SPointer -> BotTyp -> Maybe BotTyp

transpose_bw_src :: FContext -> String -> String -> Sstate SPointer -> BotSrc -> Maybe (NESet BotSrc)

transpose_bw_base :: p1 -> String -> String -> p2 -> PointerBase -> PointerBase

mk_NE_sources :: Ord a => NESet (Maybe (NESet a)) -> Maybe (NESet a)

read_sp :: FContext -> SStatePart -> State (Sstate SPointer, VCS) SPointer

data FunctionType

Constructors

AnalyzedInternalFunction (Sstate SPointer)

ExternalFunction

AnalyzedInternalFunctionTerminates

AnalyzedInternalFunctionUnknown

get_function_type :: FContext -> Instruction -> [Char] -> FunctionType

call :: FContext -> Instruction -> State (Sstate SPointer, VCS) ()

Executes semantics for external functions.

unknown_internal_function :: Monad m => p1 -> p2 -> m ()

jump :: FContext -> Instruction -> State (Sstate SPointer, VCS) ()

ctry_jump_targets :: FContext -> SPointer -> Maybe (Set ResolvedJumpTarget)

Orphan instances

SymbolicExecutable FContext SPointer

Instance details

Methods

sseparate :: FContext -> String -> SPointer -> SPointer -> SPointer -> SPointer -> Bool

senclosed :: FContext -> SPointer -> SPointer -> SPointer -> SPointer -> Bool

salias :: FContext -> SPointer -> SPointer -> SPointer -> SPointer -> Bool

ssensitive :: FContext -> SPointer -> SPointer -> SPointer -> Bool

sread_from_ro_data :: FContext -> SPointer -> SPointer -> Maybe SPointer

smk_mem_addresses :: FContext -> SPointer -> Set SPointer

sjoin :: Foldable t => FContext -> String -> t SPointer -> SPointer

swiden :: FContext -> String -> SPointer -> SPointer

top :: FContext -> SPointer

ssemantics :: FContext -> SymbolicOperation SPointer -> SPointer

sflg_semantics :: FContext -> SPointer -> Instruction -> FlagStatus -> FlagStatus

simmediate :: Integral i => FContext -> i -> SPointer

mk_svalue :: FContext -> SimpleExpr -> SPointer

mk_smem_value :: FContext -> SPointer -> SPointer -> SPointer

scall :: FContext -> Instruction -> State (Sstate SPointer, VCS) ()

sjump :: FContext -> Instruction -> State (Sstate SPointer, VCS) ()

stry_jump_targets :: FContext -> SPointer -> Maybe (Set ResolvedJumpTarget)

stry_immediate :: FContext -> SPointer -> Maybe Word64

stry_deterministic :: FContext -> SPointer -> Maybe SimpleExpr

spointer_to_exprs :: FContext -> SPointer -> Set SimpleExpr

saddress_has_instruction :: FContext -> SPointer -> Word64 -> Bool

Propagator FContext Predicate

Instance details

Methods

tau :: FContext -> [Instruction] -> Maybe [Instruction] -> Predicate -> (Predicate, Set VerificationCondition)

join :: FContext -> Predicate -> Predicate -> Predicate

implies :: FContext -> Predicate -> Predicate -> Bool