SymbolicExecution

Description

Documentation

tau_block

Arguments

:: FContext
The context

-> [Instr]
The instructions of the basic block

-> Maybe [Instr]
Optionally, the instructions of the next block if symbolically executing an edge in a CFG.

-> Pred
The predicate to be transformed

-> (Pred, VCS)

Do predicate transformation over a basic block in a CFG. Given an edge in the CFG from none block to another, perform predicate transformation. Parameter insts' is needed to set the flags properly. If Nothing is supplied, the flags are overapproximatively set to None.

init_pred

Arguments

:: FContext
The current context

-> Invariants
The currently available invariants

-> Set (NodeInfo, Pred)
The currently known postconditions

-> Pred

The initial predicate.

gather_stateparts

Arguments

:: FInit
The function initialisation

-> Invariants
The currently available invariants

-> Set (NodeInfo, Pred)
The currently known postconditions

-> Set StatePart

Given the currently known invariants and postconditions, gather all stateparts occurring in the current function.

invariant_to_finit :: FContext -> Pred -> FInit

Convert the current invariant into a function initialisation

join_finit :: FContext -> FInit -> FInit -> FInit

The join between two function initialisations

get_invariant :: FContext -> Int -> Maybe Pred

Get the invariant for a given instruction address for a given function entry

Orphan instances

Propagator FContext Pred

Instance details

Methods

tau :: FContext -> [Instr] -> Maybe [Instr] -> Pred -> (Pred, Set VerificationCondition)

join :: FContext -> Pred -> Pred -> Pred

implies :: FContext -> Pred -> Pred -> Bool